server/pyproject.toml
[tool.poetry]
name = "Sever code"
version = "0.1.0"
description = "Server source code for my blog post on multi device sign in for stateless applications using redis"
authors = ["emperorsixpacks <andrewdvd10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
redis = "^5.1.0"
fastapi = "^0.115.0"
bcrypt = "^4.2.0"
python-dotenv = "^1.0.1"
black = "^24.8.0"
uvicorn = "^0.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
